+++ src/main/java/io/mixeway/plugins/infrastructurescan/scheduler/NetworkScanScheduler.java
@@ -18,7 +18,6 @@ import io.mixeway.db.repository.NessusScanRepository;
 import io.mixeway.db.repository.ProjectRepository;
 import io.mixeway.db.repository.ScannerTypeRepository;
 import io.mixeway.plugins.infrastructurescan.service.NetworkScanClient;
-import io.mixeway.pojo.ScanHelper;
 import io.mixeway.pojo.WebAppHelper;
 import org.codehaus.jettison.json.JSONException;
 import org.slf4j.Logger;
@@ -40,14 +39,12 @@ public class NetworkScanScheduler {
 	private final List<NetworkScanClient> networkScanClients;
 	private final NetworkScanService networkScanService;
 	private final ProjectRepository projectRepository;
-	private final ScanHelper scanHelper;
 	@Autowired
 	NetworkScanScheduler(NessusScanRepository nessusScanRepository, NetworkScanService networkScanService,
 						 ScannerTypeRepository scannerTypeRepository, ProjectRepository projectRepository,
-						 WebAppHelper webAppHelper, List<NetworkScanClient> networkScanClients, ScanHelper scanHelper){
+						 WebAppHelper webAppHelper, List<NetworkScanClient> networkScanClients){
 		this.scannerTypeRepository = scannerTypeRepository;
 		this.nessusScanRepository = nessusScanRepository;
-		this.scanHelper = scanHelper;
 		this.webAppHelper = webAppHelper;
 		this.projectRepository = projectRepository;
 		this.networkScanClients = networkScanClients;
@@ -68,7 +65,7 @@ public class NetworkScanScheduler {
 							networkScanClient.loadVulnerabilities(ns);
 							ns.setRunning(false);
 							nessusScanRepository.save(ns);
-							//scanHelper.updateInterfaceState(ns,false);
+							log.info("Loaded result for {} scan of {}",ns.getNessus().getScannerType().getName(), ns.getProject().getName());
 							if (ns.getNessus().getRfwUrl() != null) {
 								networkScanService.deleteRulsFromRfw(ns);
 								log.info("RFW for scan {} is cleared - dropped traffic", ns.getProject().getName());
@@ -88,7 +85,7 @@ public class NetworkScanScheduler {
 		}
 	}
 	//Every 12h
-	@Scheduled(cron="0 15 10,21 * * *" )
+	@Scheduled(cron="#{@getNetworkCronExpresion}" )
 	public void runScheduledTest() throws JSONException, KeyManagementException, UnrecoverableKeyException, NoSuchAlgorithmException, KeyStoreException, CertificateException, IOException {
 		log.info("Starting Scheduled task for automatic test");
 		List<Project> autoInfraProjectList = projectRepository.findByAutoInfraScan(true);