+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+<a name="0.9.1"></a>
+## 0.9.1 (2020-01-14)
+
+### Info
+
+* Fixed bug with Network scan request API
+* Added possibility to put CRON expresion to DB and load if from DB
+* Fixed some minnor bugs
+
 <a name="0.9"></a>
 ## 0.9 (2019-12-07)